Fix isfoo/tofoo() warnings with Cygwin
Macros like isspace() and tolower() only have defined behavior for
unsigned char values and EOF (-1), but char may be signed
(depending on implementation), potentially passing values from -2 to
-128 to them, which Cygwin deliberately warns of.

Casting to unsigned char guarantees valid results and suppresses most of
the warnings when this is compiled with Cygwin:

    https://www.securecoding.cert.org/confluence/display/seccode/STR37-C.+Arguments+to+character+handling+functions+must+be+representable+as+an+unsigned+char

+++ libnitrohack/src/makemon.c
@@ -1204,7 +1204,7 @@ const struct permonst *rndmonst(struct level *lev)
 		rndmonst_state.mchoices[mndx] = 0;
 		if (tooweak(mndx, minmlev) || toostrong(mndx, maxmlev))
 		    continue;
-		if (upper && !isupper(def_monsyms[(int)(ptr->mlet)])) continue;
+		if (upper && !isupper((uchar)def_monsyms[(int)(ptr->mlet)])) continue;
 		if (elemlevel && wrong_elem_type(dlev, ptr)) continue;
 		if (uncommon(dlev, mndx)) continue;
 		if (In_hell(dlev) && (ptr->geno & G_NOHELL)) continue;
